refactor: split input/ files for 50-line limit compliance
- chat_spawn.rs (63→35): extract spawn task body to chat_spawn_task.rs
- chat_submit.rs (70→35): extract dispatch to chat_submit_dispatch.rs
- pr_command.rs (75→30): extract title/body to pr_title.rs, pr_body.rs
- tests.rs (124→deleted): split into tests_paste, tests_enter,
  tests_submit, tests_pr, tests_steering (one test group per file)

All files now under 50 code lines. 9/9 tests pass.

CodeTether-Provenance-ID: ctprov_a24b885ed5c84512b3de44f8bf9de30d
CodeTether-Origin: local_cli
CodeTether-Agent-Name: build
CodeTether-Session-ID: 727acf18-4fcd-4e1a-a5bc-7f1aed274952
