+++ src/tui/app/input/chat_submit.rs
@@ -1,42 +1,25 @@
 //! Chat prompt submission for the TUI.
 //!
 //! Validates the input, dispatches slash commands, pushes
-//! user messages, and spawns the provider task that runs
-//! the prompt in an optional worktree.
-//!
-//! # Examples
-//!
-//! ```ignore
-//! handle_enter_chat(&mut app, cwd, &mut session, &reg,
-//!     &bridge, &tx, &rtx).await;
-//! ```
+//! user messages, and delegates to
+//! [`super::chat_submit_dispatch::dispatch_prompt`].
 
 use std::{path::Path, sync::Arc};
 use tokio::sync::mpsc;
 
 use crate::provider::ProviderRegistry;
 use crate::session::{Session, SessionEvent};
-use crate::tui::app::{
-    commands::handle_slash_command, state::App, worker_bridge::handle_processing_started,
-};
+use crate::tui::app::commands::handle_slash_command;
+use crate::tui::app::state::App;
 use crate::tui::worker_bridge::TuiWorkerBridge;
 
-use super::chat_helpers::{no_provider_error, push_user_messages};
-use super::chat_spawn::spawn_provider_task;
+use super::chat_helpers::push_user_messages;
+use super::chat_submit_dispatch::dispatch_prompt;
 
 /// Submit the chat input to the provider.
 ///
 /// Handles slash commands, pushes user/image messages,
-/// starts processing indicators, and spawns the provider
-/// task.  If no provider is configured, an error message
-/// is shown instead.
-///
-/// # Examples
-///
-/// ```ignore
-/// handle_enter_chat(&mut app, cwd, &mut session, &reg,
-///     &bridge, &tx, &rtx).await;
-/// ```
+/// then delegates prompt preparation and dispatch.
 pub(super) async fn handle_enter_chat(
     app: &mut App,
     cwd: &Path,
@@ -67,37 +50,9 @@ pub(super) async fn handle_enter_chat(
     }
 
     push_user_messages(app, &prompt, &pending_images);
-    let effective_prompt = if let Some(prefix) = app.state.steering_prompt_prefix() {
-        format!("{prefix}\nUser request:\n{prompt}")
-    } else {
-        prompt.clone()
-    };
-    let steering_count = app.state.steering_count();
-    app.state.clear_input();
-    app.state.clear_steering();
-    handle_processing_started(app, worker_bridge).await;
-    app.state.begin_request_timing();
-    app.state.main_inflight_prompt = Some(prompt.clone());
-    app.state.status = if steering_count > 0 {
-        format!("Submitting prompt with {steering_count} steering item(s)…")
-    } else {
-        "Submitting prompt…".to_string()
-    };
-    app.state.scroll_to_bottom();
-
-    if let Some(registry) = registry {
-        spawn_provider_task(
-            app,
-            cwd,
-            session,
-            registry,
-            &effective_prompt,
-            pending_images,
-            event_tx,
-            result_tx,
-        )
-        .await;
-    } else {
-        no_provider_error(app, worker_bridge).await;
-    }
+    dispatch_prompt(
+        app, cwd, session, registry, worker_bridge,
+        &prompt, pending_images, event_tx, result_tx,
+    )
+    .await;
 }
